Modified for Python 3
 - rename project to mcerp3

 - ran 2to3 and updated for Python 3

 - adding BSD license to this repository based on the license specified
   on the PyPi package of the upstream mcerp package

 - change version to 1.0.0
