+++ README.rst
@@ -1,11 +1,11 @@
-===============================
-``mcerp`` Package Documentation
-===============================
+================================
+``mcerp3`` Package Documentation
+================================
 
 Overview
 ========
 
-``mcerp`` is a stochastic calculator for `Monte Carlo methods`_ that uses 
+``mcerp3`` is a stochastic calculator for `Monte Carlo methods`_ that uses 
 `latin-hypercube sampling`_ to perform non-order specific 
 `error propagation`_ (or uncertainty analysis). 
 
@@ -26,18 +26,10 @@ as you desire!
 What's New In This Release
 ==========================
 
-- **Significant** speed enhancements in MCERP object creation.
-
-- Updated *plotcorr* function to display correct plots when ``full``-kwarg is
-  set to ``False`` (i.e., to display only an upper triangular scatterplot
-  matrix).
-
-- Fixed a sign bug when two MCERP are compared with boolean operators >, <, 
-  etc.
+- this is a Python 3 release of the mcerp package by Abraham Lee
   
-- Improved boolean operators in general to handle MCERP objects that have
-  discrete values so that (``H(50, 5, 10)==4`` returns a meaningful probability
-  like the <, <=, >, >= operators do.
+- officially adds the 3-clause BSD licesnse text to the software
+  (this license has been specified in the mcerp PyPI package for years)  
 
 
 Main Features
@@ -65,6 +57,8 @@ Main Features
    functions. Depending on your version of Scipy, some functions might not
    work.
 
+8. Python 3 support
+
 Installation
 ============
 
@@ -84,11 +78,11 @@ scientific computing python bundles.
 How to install
 --------------
 
-You have **several easy, convenient options** to install the ``mcerp`` 
+You have **several easy, convenient options** to install the ``mcerp3`` 
 package (administrative privileges may be required)
 
-#. Simply copy the unzipped ``mcerp-XYZ`` directory to any other location that
-   python can find it and rename it ``mcerp``.
+#. Simply copy the unzipped ``mcerp3-XYZ`` directory to any other location that
+   python can find it and rename it ``mcerp3``.
     
 #. From the command-line, do one of the following:
    
@@ -99,29 +93,11 @@ package (administrative privileges may be required)
 
    b. If ``setuptools`` is installed, run::
 
-       $ [sudo] easy_install [--upgrade] mcerp
+       $ [sudo] easy_install [--upgrade] mcerp3
     
    c. If ``pip`` is installed, run::
 
-       $ [sudo] pip install [--upgrade] mcerp
-
-Python 3
---------
-
-To use this package with Python 3.x, you will need to run the ``2to3`` 
-conversion tool at the command-line using the following syntax while in the 
-unzipped ``mcerp`` directory::
-
-    $ 2to3 -w .
-    
-This should take care of the main changes required. Then, run::
-
-    $ python3 setup.py install
-
-If bugs continue to pop up, please email the author.
-
-You can also get the bleeding-edge code from `GitHub`_ (though I can't 
-promise there won't be stability issues...).
+       $ [sudo] pip install [--upgrade] mcerp3
 
 See also
 ========
@@ -132,8 +108,8 @@ See also
 Contact
 =======
 
-Please send **feature requests, bug reports, or feedback** to 
-`Abraham Lee`_.
+Please send **Python 3 related issues** to `Paul Freeman`_. Other issues should
+be referred to the original author, `Abraham Lee`_.
 
 
     
@@ -147,7 +123,8 @@ Please send **feature requests, bug reports, or feedback** to
 .. _Matplotlib: http://matplotlib.org/
 .. _scipy.stats: http://docs.scipy.org/doc/scipy/reference/stats.html
 .. _uncertainties: http://pypi.python.org/pypi/uncertainties
-.. _source code: https://github.com/tisimst/mcerp
+.. _source code: https://github.com/paul-freeman/mcerp
 .. _Abraham Lee: mailto:tisimst@gmail.com
-.. _package documentation: http://pythonhosted.org/mcerp
-.. _GitHub: http://github.com/tisimst/mcerp
+.. _Paul Freeman: mailto:paul.freeman.cs@gmail.com
+.. _package documentation: http://pythonhosted.org/mcerp3
+.. _GitHub: http://github.com/paul-freeman/mcerp
