+++ mcerp3/lhd.py
@@ -177,7 +177,7 @@ def lhd(dist=None,size=None,dims=1,form='randomized',iterations=100,
         if dim=='rows':
             data = data.T
 
-        data_rank = range(n)
+        data_rank = list(range(n))
         for i in range(data.shape[1]):
             new_data_rank = np.random.permutation(data_rank)
             vals, order = np.unique(np.hstack((data_rank, new_data_rank)), return_inverse=True)
@@ -228,7 +228,7 @@ def lhd(dist=None,size=None,dims=1,form='randomized',iterations=100,
                 
                 data = _mix(data)
             
-            print 'Optimized Distance:',d_opt
+            print('Optimized Distance:',d_opt)
             return data_opt
 
         if hasattr(dist,'__getitem__'): # if multiple distributions were input
@@ -269,9 +269,9 @@ def lhd(dist=None,size=None,dims=1,form='randomized',iterations=100,
         VIF = np.max(np.diag(inv_cor_matrix))
             
         if showcorrelations:
-            print 'Correlation Matrix:\n',cor_matrix
-            print 'Inverted Correlation Matrix:\n',inv_cor_matrix
-            print 'Variance Inflation Factor (VIF):',VIF
+            print('Correlation Matrix:\n',cor_matrix)
+            print('Inverted Correlation Matrix:\n',inv_cor_matrix)
+            print('Variance Inflation Factor (VIF):',VIF)
 #        elif VIF >= 1:
 #            print 'WARNING: Variance Inflation Factor (%5.3f) indicates that there'%(VIF)
 #            print 'may be some undesirably large pairwise correlations present.'
@@ -281,16 +281,16 @@ def lhd(dist=None,size=None,dims=1,form='randomized',iterations=100,
 if __name__=='__main__':
     # test single distribution
     d0 = ss.uniform(loc=-1,scale=2) # uniform distribution,low=-1, width=2
-    print lhd(dist=d0,size=5)
+    print(lhd(dist=d0,size=5))
     
     # test single distribution for multiple variables
     d1 = ss.norm(loc=0,scale=1) # normal distribution, mean=0, stdev=1
-    print lhd(dist=d1,size=7,dims=5)
+    print(lhd(dist=d1,size=7,dims=5))
     
     # test multiple distributions
     d2 = ss.beta(2,5) # beta distribution, alpha=2, beta=5
     d3 = ss.expon(scale=1/1.5) # exponential distribution, lambda=1.5
-    print lhd(dist=(d1,d2,d3),size=6)
+    print(lhd(dist=(d1,d2,d3),size=6))
     
     rand_lhs = lhd(dist=(d0,d1,d2,d3),size=100)
     spac_lhs = lhd(dist=(d0,d1,d2,d3),size=100,form='spacefilling',
@@ -300,8 +300,8 @@ if __name__=='__main__':
         from scatterplot_matrix import scatterplot_matrix as spm
         import matplotlib.pyplot as plt
     except ImportError:
-        print rand_lhs
-        print spac_lhs
+        print(rand_lhs)
+        print(spac_lhs)
     else:
         names = ['U(-1,1)','N(0,1)','Beta(2,5)','Exp(1.5)']
         spm(rand_lhs.T,names=names)
