+++ mcerp3/umath.py
@@ -1,6 +1,6 @@
 """
 ================================================================================
-mcerp: Real-time latin-hypercube-sampling-based Monte Carlo Error Propagation
+mcerp3: Real-time latin-hypercube-sampling-based Monte Carlo Error Propagation
 ================================================================================
 
 Generalizes mathematical operators that work on numeric objects (from the math
@@ -8,8 +8,10 @@ module or numpy) compatible with objects with uncertainty distributions
 
 Author: Abraham Lee
 Copyright: 2013
+
+Updates to Python 3 by Paul Freeman, 2018
 """
-from mcerp import UncertainFunction, to_uncertain_func
+from mcerp3 import UncertainFunction, to_uncertain_func
 import numpy as np
 
 __author__ = 'Abraham Lee'
