+++ revision_history.txt
@@ -1,3 +1,12 @@
+mcerp3-1.0.0
+============
+
+- updated for Python 3
+
+- changed instances where np.round() calls were expected to return an integer
+
+- added 3-clause BSD License text
+
 mcerp-0.9.2
 ===========
